Move back to old folder structure
