+++ asteroid.rb
